likes service: switch to grpc

+++ services/likes/models.go
@@ -5,13 +5,13 @@ import (
 )
 
 type PostLike struct {
-	PostID    uint `gorm:"primary_key auto_increment:false"`
-	UserID    uint `gorm:"primary_key auto_increment:false"`
+	PostID    uint64 `gorm:"primary_key auto_increment:false"`
+	UserID    uint64 `gorm:"primary_key auto_increment:false"`
 	CreatedAt time.Time
 }
 
 type CommentLike struct {
-	CommentID uint `gorm:"primary_key auto_increment:false"`
-	UserID    uint `gorm:"primary_key auto_increment:false"`
+	CommentID uint64 `gorm:"primary_key auto_increment:false"`
+	UserID    uint64 `gorm:"primary_key auto_increment:false"`
 	CreatedAt time.Time
 }
